feat: restores winview in every window
Originally, when #1 was requested, I made a window state restoration.
But it only worked in a current window and created a bug when operating
in remote windows. Now this is fixed and works with no errors or bugs.

When jumping to another window, its state will be saved to then restore
when leaving it. The same with a current window.